add Product, Transaction Service Interface
